refactor(calendar): finir la décomposition d'UniversalCalendar (<300 lignes) (#107)
Poursuit #28bis jusqu'à faire passer UniversalCalendar.vue de 1154 à 230 lignes.
L'orchestrateur ne fait plus que de la composition + l'état partagé (vue, date,
événement sélectionné, filtres) ; le rendu est éclaté en sous-composants :

- CalendarNavigation.vue : barre de navigation (mois/vues/actualiser), props +
  émetteurs, aucun accès direct à FullCalendar.
- CalendarGrid.vue : encapsule FullCalendar (plugins, options, sync impérative des
  événements) ; expose prev/next/today/changeView et émet event-click/dates-set/
  view-changed. Porte les styles .fc (light + dark).
- CalendarLegend.vue : légende, purement présentationnel.
- CalendarFilters.vue : rapatrie ses styles dark restés orphelins dans le parent
  (#28bis) et retire le .card dupliqué.

CSS « chrome » partagé (carte de base + conteneur + dark de base) centralisé dans
le parent via :deep() / sélecteurs globaux, sans duplication. Suppression d'un
import useRouter mort.

Contrat public inchangé (props userRole/userId, emit event-action, expose
refreshEvents) : les 3 vues consommatrices ne bougent pas. Tests 277/277, build
vert, vérification visuelle light/dark + changement de vue OK.
